requirements.txt
streamlit
torch
torchvision
pillow
numpy
opencv-python
psycopg2-binary
streamlit-drawable-canvas